switch to more standard typer configuration
previous implementation aimed to mirror fire behaviors
