+++ ensembl_genes/ensembl_genes.py
@@ -599,63 +599,3 @@ class Ensembl_Gene_Catalog_Writer(Ensembl_Gene_Queries):
             output_path=self.output_directory.joinpath(notebook),
             parameters={"species": self.species.common_name, "release": self.release},
         )
-
-
-class Commands:
-    @staticmethod
-    def export_datasets(species: str = "human", release: str = "latest") -> None:
-        """Export datasets to output directory."""
-        ensgc = Ensembl_Gene_Catalog_Writer(species=species, release=release)
-        logging.info(
-            f"exporting ensembl genes to {ensgc.output_directory}: version {ensgc.release}"
-        )
-        logging.info(f"connection_url: {ensgc.connection_url}")
-        ensgc.export_datasets()
-
-    @staticmethod
-    def export_notebooks(species: str = "human", release: str = "latest") -> None:
-        """Execute notebooks using papermill and save results in output directory."""
-        logging.info("Executing notebooks with papermill")
-        ensgc = Ensembl_Gene_Catalog_Writer(species=species, release=release)
-        ensgc.export_notebooks()
-
-    @classmethod
-    def export_all(cls, species: str = "human", release: str = "latest") -> None:
-        """Export datasets and then notebooks."""
-        cls.export_datasets(species=species, release=release)
-        cls.export_notebooks(species=species, release=release)
-
-    @staticmethod
-    def get_ensembl_database(species: str = "human", release: str = "latest") -> str:
-        """
-        Return ensembl core database string for a species & release.
-        See <http://ftp.ensembl.org/pub/current_mysql/> for a list of current ensembl databases.
-        """
-        ensgc = Ensembl_Gene_Queries(species=species, release=release)
-        return ensgc.database
-
-    @classmethod
-    def command(cls) -> None:
-        """
-        Run like `poetry run ensembl_genes`
-        """
-        import typer
-
-        logging.basicConfig()
-        logging.getLogger().setLevel(logging.INFO)
-        commands = {
-            "datasets": cls.export_datasets,
-            "notebooks": cls.export_notebooks,
-            "all": cls.export_all,
-            "ensembl_release": check_ensembl_release,
-            "ensembl_database": cls.get_ensembl_database,
-        }
-
-        def print_result(result: str) -> None:
-            if result is not None:
-                typer.echo(result)
-
-        cli = typer.Typer(result_callback=print_result)
-        for command_name, command_fxn in commands.items():
-            cli.command(name=command_name)(command_fxn)  # type: ignore [type-var]
-        cli()
